Remove the 'needCamelSnakeConversion' flag
This was an artifact of how we used to do things a *long* time ago and
needed the conversion to be exactly reversible.  We no longer need that,
and the rewritten casing converters (from a couple months ago) are
idempotent and much more robust.

Before+after of all generated examples shows zero changes in any
generated code.