Add deduping support to ProtocolParameters
Previously, the 'needSnakeCamelConversion' flag would accidentally cause
some cases with duplicate argument names to actually work out and not
end up with duplicates.  Since I removed that flag, we can now have
duplicates, so we need a deduping method.  This uses the same method as
in ScalaParameter: if there are any conflicts, we instead use the raw
property name (escaped) for the conflicting properties.